Collapses LogAccess and LogAccessHttp into LogAccess
* make ProxyError::str inline so traffic_logstats/logcat doesn't need
  to pull in an even larger number of dependencies.
