+++ proxy/logging/LogAccessTest.h
@@ -46,9 +46,8 @@ public:
   //
   // client -> proxy fields
   //
-  virtual int marshal_client_host_ip(char *);        // INT
-  virtual int marshal_client_auth_user_name(char *); // STR
-  // marshal_client_req_timestamp_sec is non-virtual!
+  virtual int marshal_client_host_ip(char *);            // INT
+  virtual int marshal_client_auth_user_name(char *);     // STR
   virtual int marshal_client_req_text(char *);           // STR
   virtual int marshal_client_req_http_method(char *);    // INT
   virtual int marshal_client_req_url(char *);            // STR
